Robots file reader repaired
The robots file reader now remembers a robots.txt file for each
distinct hostname found during navigation and reuses it across
all threads. Instead of handing around a function handle a thread-
safe object is used to keep track of the them, making the code
a little easier to read.

Reenabled the negative test which was disabled.

Fixes #62